split draw
